pyproject.toml
[tool.poetry]
name = "github-event-monitor"
version = "0.1.0"
description = "A system to monitor GitHub events and provide metrics using Medallion Architecture"
authors = ["Berkay Zuhre"]
readme = "README.md"
packages = [{ include = "github_event_monitor" }]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
uvicorn = "^0.23.2"
sqlalchemy = "^2.0.23"
aiosqlite = "^0.19.0"
aiohttp = "^3.8.6"
apscheduler = "^3.10.4"
dash = "^2.14.1"
pandas = "^2.1.2"
plotly = "^5.18.0"
pydantic = "^2.4.2"
python-dateutil = "^2.8.2"
python-dotenv = "^1.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
black = "^23.10.1"
isort = "^5.12.0"
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
